the core concepts: core minimal interface, with extension interfaces
